Got blueprints working, updated url references
