+++ dashboard/main/__init__.py
@@ -1,5 +1,5 @@
 from flask import Blueprint
 
-auth_bp = Blueprint('auth', __name__)
+main_bp = Blueprint('main', __name__)
 
 from . import views
