+++ dashboard/templates/404.html
@@ -3,5 +3,5 @@
 
 {% block content %}
   <h1>File Not Found</h1>
-  <p><a href="{{ url_for('index') }}">Back</a></p>
+  <p><a href="{{ url_for('main.index') }}">Back</a></p>
 {% endblock %}
