+++ dashboard/templates/500.html
@@ -4,5 +4,5 @@
 {% block content %}
   <h1>An unexpected error has occurred</h1>
   <p>The administrator has been notified. Sorry for the inconvenience!</p>
-  <p><a href="{{ url_for('index') }}">Back</a></p>
+  <p><a href="{{ url_for('main.index') }}">Back</a></p>
 {% endblock %}
