+++ dashboard/templates/timepoint/sub_header.html
@@ -1,7 +1,7 @@
 <div class="row">
   <div class="col-xs-4">
     <div class="btn-group" role="group">
-      <a class="btn btn-primary" href="{{ url_for('study', study_id=study_id) }}">
+      <a class="btn btn-primary" href="{{ url_for('main.study', study_id=study_id) }}">
         <span class="glyphicon glyphicon-arrow-left"></span>
         <span class="button-text">Return to {{ study_id }}</span>
       </a>
