+++ dashboard/users/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+user_bp = Blueprint('users', __name__, template_folder='templates')
+
+from . import views
