+++ dashboard/users/templates/account_request.html
@@ -3,11 +3,11 @@
 
 {% block content %}
   <div class="container">
-    <form action="{{ url_for('new_account') }}" method="post"
+    <form action="{{ url_for('users.new_account') }}" method="post"
         name="new_account">
       {{ form.hidden_tag() }}
       <h2>New User Details</h2>
-      {% include 'users/demographics_snip.html' %}
+      {% include 'snips/demographics.html' %}
       <input class="btn btn-primary" type="submit" value="Submit"
           style="margin-top: 10px;">
     </form>
