+++ dashboard/users/templates/snips/permissions.html
