+++ dashboard/view_utils.py
@@ -7,72 +7,14 @@ import logging
 from functools import wraps
 
 from urllib.parse import urlparse, urljoin
-from flask import flash, url_for, current_user, request, redirect
+from flask_login import current_user
+from flask import flash, url_for, request, redirect
 from werkzeug.routing import RequestRedirect
 
 from .models import Study, Timepoint, Scan, RedcapRecord, User
-from .forms import UserForm, UserAdminForm
 from .utils import create_issue as make_issue
 
 logger = logging.getLogger(__name__)
-
-
-def get_user_form(user, current_user):
-    if not current_user.dashboard_admin:
-        form = UserForm(obj=user)
-        form.account.data = user.username
-        return form
-    form = get_admin_user_form(user)
-    return form
-
-
-def get_admin_user_form(user):
-    form = UserAdminForm(obj=user)
-    form.account.data = user.username
-    choices = populate_disabled_sites(user)
-    form.add_access.choices = choices
-    return form
-
-
-def populate_disabled_sites(user):
-    disabled = user.get_disabled_sites()
-    choices = []
-    for study in disabled:
-        if len(disabled[study]) > 1:
-            choices.append((study, study + " - ALL"))
-        for site in disabled[study]:
-            choices.append((study + "-" + site, study + " - " + site))
-    return choices
-
-
-def parse_enabled_sites(new_access):
-    """Parses the UserAdminForm add_access field into a dictionary
-
-    Args:
-        new_access (:obj:`list`): A list of 'STUDY-SITE' and 'STUDY' strings
-        like the sort returned by UserAdminForm.add_access.data
-
-    Returns:
-        :obj:`dict`: A dictionary mapping each study to a list of sites to
-        enable. The empty list indicates 'all sites'.
-    """
-    enabled = {}
-    for option in new_access:
-        fields = option.split('-')
-        study = fields[0]
-        try:
-            site = fields[1]
-        except IndexError:
-            # Grant user access to all sites for this study
-            enabled[study] = []
-        else:
-            # Grant user access to subset of sites for this study
-            if enabled.get(study) == []:
-                # If the user is already being given global study access
-                # don't accidentally restrict them to a subset of sites
-                continue
-            enabled.setdefault(study, []).append(site)
-    return enabled
 
 
 def report_form_errors(form):
@@ -125,12 +67,12 @@ def get_timepoint(study_id, timepoint_id, current_user):
 
     if timepoint is None:
         flash("Timepoint {} does not exist".format(timepoint_id))
-        raise RequestRedirect(url_for("index"))
+        raise RequestRedirect(url_for("main.index"))
 
     if (not current_user.has_study_access(study_id, timepoint.site_id) or
             not timepoint.belongs_to(study_id)):
         flash("Not authorised to view {}".format(timepoint_id))
-        raise RequestRedirect(url_for("index"))
+        raise RequestRedirect(url_for("main.index"))
 
     return timepoint
 
@@ -147,7 +89,7 @@ def get_session(timepoint, session_num, fail_url):
 
 def get_scan(scan_id, study_id, current_user, fail_url=None):
     if not fail_url:
-        fail_url = url_for('index')
+        fail_url = url_for('main.index')
 
     scan = Scan.query.get(scan_id)
 
@@ -168,7 +110,7 @@ def get_scan(scan_id, study_id, current_user, fail_url=None):
 
 def get_redcap_record(record_id, fail_url=None):
     if not fail_url:
-        fail_url = url_for('index')
+        fail_url = url_for('main.index')
 
     record = RedcapRecord.query.get(record_id)
 
@@ -217,7 +159,7 @@ def prev_url():
     """
     if request.referrer and is_safe_url(request.referrer):
         return request.referrer
-    return url_for('index')
+    return url_for('main.index')
 
 
 def is_safe_url(target):
